di: move Extension to appropriate namespace, instantiate explicitely
