+++ DependencyInjection/Extension/CoreSphereConsoleExtension.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace CoreSphere\ConsoleBundle\DependencyInjection;
+namespace CoreSphere\ConsoleBundle\DependencyInjection\Extension;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
@@ -24,7 +24,7 @@ final class CoreSphereConsoleExtension extends Extension implements PrependExten
      */
     public function load(array $configs, ContainerBuilder $container)
     {
-        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../../Resources/config'));
         $loader->load('services.yml');
     }
